@@ -6,6 +6,7 @@
     "dependencies": {
         "@babel/plugin-proposal-class-properties": "^7.12.1",
         "@babel/plugin-transform-runtime": "^7.12.10",
+        "axios": "^0.21.1",
         "cors": "^2.8.5",
         "express": "^4.17.1",
         "jquery": "^3.5.1",
@@ -18,19 +19,28 @@
         "react-router-dom": "^5.2.0",
         "react-toastify": "^7.0.2",
         "styled-components": "^5.2.1",
-        "webpack": "^4.44.2",
+        "webpack": "^4.46.0",
         "webpack-cli": "^3.3.12"
     },
     "devDependencies": {
         "@babel/core": "^7.12.10",
         "@babel/preset-env": "^7.12.11",
         "@babel/preset-react": "^7.12.10",
+        "@types/express": "^4.17.13",
+        "@types/mongodb": "^4.0.7",
+        "@types/node": "^17.0.7",
+        "@types/react-dom": "^17.0.11",
+        "@types/react-router-dom": "^5.3.2",
+        "@types/styled-components": "^5.1.19",
         "babel-loader": "^8.2.2",
         "css-loader": "^5.0.1",
-        "nodemon": "^2.0.4"
+        "nodemon": "^2.0.4",
+        "ts-loader": "^8.2.0",
+        "typescript": "^4.5.4"
     },
     "scripts": {
-        "start": "npx webpack && nodemon index.js",
+        "start": "npx webpack && nodemon index.ts",
+        "server": "webpack --config webpack-server.config.js && nodemon dist/server.js",
         "build": "webpack",
         "watch": "webpack --watch"
     },
