@@ -1,5 +1,5 @@
-const express = require("express");
-const fs = require("fs");
+import express from "express";
+import fs from "fs";
 
 const router = express.Router();
 
@@ -14,4 +14,4 @@ router.get("/:name", async (req, res) => {
     // #TODO
 });
 
-module.exports = router;
+export default router;
