@@ -1,5 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./lobby/app.jsx";
+import App from "./lobby/app";
 
 ReactDOM.render(<App />, document.getElementById("main"));
